Ligeros cambios en la sección contacto

+++ src/app/pages/contacto/components/show-messages/show-messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ContactMessage } from '../../services/contact-message';
 import { ContactService } from '../../services/contact.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-show-messages',
@@ -31,7 +32,24 @@ export class ShowMessagesComponent implements OnInit {
   }
 
   deleteMessage(event, mensaje: ContactMessage) {
-    this.contactService.deleteMessage(mensaje);
+    Swal.fire({
+      title: '¿Estas seguro?',
+      text: '¡Este cambio no se puede revertir!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, ¡eliminar!'
+    }).then((result) => {
+      if (result.value) {
+        this.contactService.deleteMessage(mensaje);
+        Swal.fire(
+          '¡Eliminado!',
+          'Se eliminó con éxito tu mensaje.',
+          'success'
+        );
+      }
+    });
   }
 
 }
